% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apply.factor.grouping}
\alias{apply.factor.grouping}
\title{Apply groupings to new data}
\usage{
apply.factor.grouping(data, variables, new.data, other.factor = "other")
}
\arguments{
\item{data}{data with already grouped data}

\item{variables}{factors to analyze}

\item{new.data}{new data frame with factor variables to convert}

\item{other.factor}{name of the "other" group}
}
\value{
new data frame with added columns named group.X (where X is variable name)
}
\description{
Groups factor values according to mapping generated as an output of a function group.factor. Thus the same grouping can be applied to new datasets.
Now it uses dataset itself as a mapping, which is not very efficient. In future releases a proper mapping will be generated and used.
}
\examples{
new.df <- apply.factor.grouping(df, c('var1','var2','var3'), new.df)
}

