% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{classify.columns}
\alias{classify.columns}
\title{Get variables class information}
\usage{
classify.columns(data, y.name = "target", max.cardinality = 6)
}
\arguments{
\item{data}{data frame with variables of interest}

\item{y.name}{target variable}

\item{max.cardinality}{maximum levels of numeric variable to be marked for conversion to factor}

\item{convert}{if not set - return information about variables; if set - return converted data frame and print information}
}
\value{
data frame with column types and statistics
}
\description{
Creates data frame with info about dataset columns including proposed conversions and optionally convert
}
\examples{
df.cols <- classify.columns(df)
}

