% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.woe}
\alias{compute.woe}
\title{Create cross-tabulation between factor and binary target}
\usage{
compute.woe(factor, target)
}
\arguments{
\item{factor}{factor variable}

\item{target}{binary target variable}
}
\value{
Data frame with the number of cases of each class for every factor value, lift and woe.
}
\description{
Creates a table with stats about factor values. Also computes Lift and WOEs. Print totals. Sorts result by lift.
}
\examples{
compute.woe(df$var, df$target)
}

