% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert.columns}
\alias{convert.columns}
\title{Convert columns}
\usage{
convert.columns(df, data)
}
\arguments{
\item{df}{- data frame with information about variables (may be result of classify.columns function)}

\item{data}{- data frame to convert}
}
\value{
data frame with mapping (variable value) -> (variable group)
}
\description{
Converts dataset columns according to proposed mapping
}
\examples{
df <- convert.columns(classify.columns(df), df)

}

