% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create.woe}
\alias{create.woe}
\title{Create WOE mapping}
\usage{
create.woe(x, vars = colnames(x), y)
}
\arguments{
\item{x}{data.frame with varaibles}

\item{vars}{variables of interest (must be factors!)}

\item{y}{target variable}
}
\value{
data.frame with WOE mapping (value->woe) that can be further applied to new datasets
}
\description{
Computes WOE
 If one group is empty, woe := 2 * max(WOE) (or 2 * min(WOE))
}
\examples{
create.woe(df, c('var1','var2','var3'), df$target)
}

