% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cv.glm}
\alias{cv.glm}
\title{Compute average AUC with cross-validation}
\usage{
cv.glm(formula, data, target.var, n, fit.type = "glm", train.rate = 0.7)
}
\arguments{
\item{formula}{formula to fit regression on}

\item{data}{input data.frame}

\item{target.var}{target variable name}

\item{n}{number of repetitions}

\item{fit.type}{Type of model to use (GLM or RandomForest)}

\item{train.rate}{fraction of data to set as training (default=0.7)}
}
\description{
Repeats random split, fits GLM model on train dataset and computes AUC on test dataset and calculates average AUC
In further releases will be replaced with more efficient cross-validation techniques
}

