% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate.date.features}
\alias{generate.date.features}
\title{Generate date features}
\usage{
generate.date.features(data, date.variables, overwrite = FALSE)
}
\arguments{
\item{data}{data frame of interest}

\item{date.variables}{vector of date variable names}

\item{overwrite}{let function overwrite already existing variables with the same names}
}
\value{
data frame with new generated variables
}
\description{
Generates features based on columns of date type. Creates year, month and day features both numeric and factor.
}
\examples{
df <- generate.date.features(df, c('date1','date2'))
}

