% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.auc.importance}
\alias{get.auc.importance}
\title{Calculate AUC importance of each attribute}
\usage{
get.auc.importance(x, y)
}
\arguments{
\item{x}{data frame with variables of interest (all must be numerics)}

\item{y}{target variable}
}
\value{
data frame with variable names sorted by AUC statistics
}
\description{
Creates a data frame with calculated AUC difference between good class and bad class for each variable
}

