% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.chisq.importance}
\alias{get.chisq.importance}
\title{Calculate Chi-squared importance of each attribute}
\usage{
get.chisq.importance(x, y)
}
\arguments{
\item{x}{data frame with variables of interest (all must be factors)}

\item{y}{target variable}
}
\value{
data frame with variable names sorted by chi-squared statistics
}
\description{
Creates a data frame with calculated chi-squared difference between good class and bad class for each factor
}

