% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.date.columns}
\alias{get.date.columns}
\title{Get names of all date columns}
\usage{
get.date.columns(data)
}
\arguments{
\item{data}{data.frame of interest}
}
\value{
vector of date variable names
}
\description{
Get names of all date columns
}
\examples{
df.numerics <- get.date.columns(df)
}

