% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.factor.columns}
\alias{get.factor.columns}
\title{Get names of all factor columns}
\usage{
get.factor.columns(data)
}
\arguments{
\item{data}{data frame of interest}
}
\value{
vector of factor variable names
}
\description{
Get names of all factor columns
}
\examples{
df.numerics <- get.factor.columns(df)
}

