% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.numeric.columns}
\alias{get.numeric.columns}
\title{Get names of all numeric columns}
\usage{
get.numeric.columns(data)
}
\arguments{
\item{data}{data frame of interest}
}
\value{
vector of numeric variable names
}
\description{
Get names of all numeric columns
}
\examples{
df.numerics <- get.numeric.columns(df)
}

