% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.proposed.conversions}
\alias{get.proposed.conversions}
\title{Get proposed type conversions}
\usage{
get.proposed.conversions(df)
}
\arguments{
\item{df}{output of classify.columns function}
}
\value{
data frame with variables to convert
}
\description{
Selects conversions suggested by classify.columns function
}
\examples{
get.proposed.conversions(classify.columns(df))
}

