% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.rf.importance}
\alias{get.rf.importance}
\title{Calculate RandomForest importance of each attribute}
\usage{
get.rf.importance(x, y)
}
\arguments{
\item{x}{data frame with variables of interest}

\item{y}{target variable}
}
\value{
data frame with variable names sorted by RandomForest statistics
}
\description{
Creates a data frame with calculated RandomForest importance for each variable
}

