% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.rf.subset.quality}
\alias{get.rf.subset.quality}
\title{Beta}
\usage{
get.rf.subset.quality(data, sorted.variables, y.name, analyze.top.n = 50,
  RF.trees = 100, times = 1, trainSplit = 0.7, logging = "sparse")
}
\description{
Beta
}

