% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{group.factor}
\alias{group.factor}
\title{Put rare factor values into separate group}
\usage{
group.factor(x, top.values = 10, min.frequency = NA)
}
\arguments{
\item{x}{data vector}

\item{top.values}{number of bins to create (plus "other" group)}

\item{min.frequency}{if factor value is less requent than this value, then put it to "other" group}
}
\value{
data frame with mapping (variable value) -> (variable group)
}
\description{
Sets rare values of a factor to new value "other". Thus limits the number of factor levels.
}
\examples{
df$var <- group.factor(df$var, top.values=10)
df$var <- group.factor(df$var, min.frequency=100)
}

