% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{optimal.factor.grouping}
\alias{optimal.factor.grouping}
\title{Optimal factor cutoff (Beta)}
\usage{
optimal.factor.grouping(x, y, fit.type = "glm", max.levels = 30)
}
\description{
Optimal factor cutoff (Beta)
}

