% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set.missing.factors.to.NA}
\alias{set.missing.factors.to.NA}
\title{Replace empty factor values with "NA"}
\usage{
set.missing.factors.to.NA(data, variables, logging = FALSE)
}
\arguments{
\item{data}{data frame with variables of interest}

\item{variables}{vector of factors}

\item{logging}{flag to write detailed stats in standard output (default=FALSE)}
}
\value{
data frame with replaced missing values for selected variables
}
\description{
Replaces blank factor values with "NA". Recreates factor variable with new level.
}
\examples{
set.missing.factors.to.NA(df, c('var1','var2','var3'), logging=TRUE)
}

