% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set.missing.to.mean}
\alias{set.missing.to.mean}
\title{Replace numeric missing values with mean}
\usage{
set.missing.to.mean(data, variables)
}
\arguments{
\item{data}{- data frame with variables of interest}

\item{variables}{- vector of numeric variables to replace missing values}
}
\value{
data frame with replaced missing values for selected variables
}
\description{
Replaces missing (NA) values with column mean. Only applicable for numerics.
}

