% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set.missing.to.prediction}
\alias{set.missing.to.prediction}
\title{Predict missing numeric values (beta)}
\usage{
set.missing.to.prediction(x)
}
\description{
Uses regression to predict most probable value
Was not tested yet
}

