% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set.missing.to.random}
\alias{set.missing.to.random}
\title{Set missing to random}
\usage{
set.missing.to.random(data, variables)
}
\arguments{
\item{data}{data frame of interest}

\item{variables}{variables with possibly missing values}
}
\value{
data frame with missing values replaced by values from random sample
}
\description{
Sets missing values to values from a random sample, which repeats distribution of non-missing values
}

