% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set.missing.to.zero}
\alias{set.missing.to.zero}
\title{Replace numeric missing values with zero}
\usage{
set.missing.to.zero(data, variables)
}
\arguments{
\item{data}{data frame with variables of interest}

\item{variables}{vector of numeric variables to replace missing values}
}
\value{
data frame with replaced missing values for selected variables
}
\description{
Replaces missing (NA) values with zero. Only applicable for numerics.
}

