% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{woe.apply}
\alias{woe.apply}
\title{Apply WOE mapping to new data}
\usage{
woe.apply(x, woe.map)
}
\arguments{
\item{x}{data.frame with factor values to replace}

\item{woe.map}{generated with create.woe (or manually) mapping}
}
\description{
Replaces factor values with corresponding WOE scores
}
\examples{
woe.apply(df, woe.map)
}

